# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShortURL Pro API",
    "description": "ShortURL Pro 是一个基于 Spring Boot 的短链接服务系统，支持链接缩短、访问统计、用户管理等功能。\n\n## 核心功能\n- 🔗 短链接生成：将长URL转换为短码\n- 🔄 302跳转：根据短码重定向到原始链接\n- 📊 访问统计：实时记录点击次数\n- 🔐 用户管理：基于Spring Security的权限控制\n- 🛠️ 管理功能：启用/禁用短链接状态\n\n## 访问权限\n- **公开接口**：短链接生成、跳转、演示页\n- **认证接口**：用户登录、登出\n- **管理员接口**：需要 ROLE_ADMIN 角色",
    "version": "1.0.0",
    "contact": {
      "name": "ShortURL Pro Team",
      "email": "support@shorturlpro.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    },
    {
      "url": "https://api.shorturlpro.com",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/api/short-url": {
      "post": {
        "tags": ["管理接口"],
        "summary": "创建短链接",
        "description": "创建新的短链接记录\n\n**权限要求**：需要认证",
        "operationId": "createShortUrl",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/short-url/{id}": {
      "put": {
        "tags": ["管理接口"],
        "summary": "更新短链接",
        "description": "根据ID更新短链接信息\n\n**权限要求**：需要认证",
        "operationId": "updateShortUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "短链接ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShortUrlCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShortUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "短链接不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["管理接口"],
        "summary": "删除短链接",
        "description": "根据ID删除短链接记录\n\n**权限要求**：需要认证",
        "operationId": "deleteShortUrl",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "短链接ID"
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "短链接不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin": {
      "get": {
        "tags": ["管理接口"],
        "summary": "管理后台首页",
        "description": "管理后台主页\n\n**权限要求**：需要 ROLE_ADMIN 角色",
        "operationId": "adminDashboard",
        "responses": {
          "200": {
            "description": "管理页面",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "未认证"
          },
          "403": {
            "description": "权限不足"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ShortUrlCreateRequest": {
        "type": "object",
        "required": ["originalUrl", "name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "短链接名称",
            "example": "我的网站",
            "maxLength": 100
          },
          "originalUrl": {
            "type": "string",
            "format": "uri",
            "description": "原始长链接",
            "example": "https://www.example.com"
          },
          "appId": {
            "type": "string",
            "description": "应用标识（可选）",
            "example": "myapp",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "过期时间（可选）",
            "example": "2026-12-31T23:59:59Z",
            "nullable": true
          }
        }
      },
...
      "ShortUrlResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "短链接名称",
            "example": "我的网站"
          },
          "originalUrl": {
            "type": "string",
            "format": "uri",
            "description": "原始长链接",
            "example": "https://www.example.com"
          },
          "shortCode": {
            "type": "string",
            "description": "短码",
            "example": "abc123xyz"
          },
          "status": {
            "$ref": "#/components/schemas/ShortUrlStatus"
          },
          "clickCount": {
            "type": "integer",
            "format": "int64",
            "description": "点击次数",
            "example": 100
          },
          "appId": {
            "type": "string",
            "description": "应用标识",
            "example": "myapp",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "过期时间",
            "example": "2026-12-31T23:59:59Z",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间",
            "example": "2024-01-01T12:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间",
            "example": "2024-01-01T12:00:00Z"
          }
        }
      },
      "ShortUrlStatus": {
        "type": "string",
        "description": "短链接状态",
        "enum": ["ENABLED", "DISABLED"],
        "example": "ENABLED"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "错误发生时间"
          },
          "status": {
            "type": "integer",
            "description": "HTTP状态码"
          },
          "error": {
            "type": "string",
            "description": "错误类型"
          },
          "message": {
            "type": "string",
            "description": "错误消息"
          },
          "path": {
            "type": "string",
            "description": "请求路径"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
}